// wesignsigner-client-app-DEV/src/app/enums/store-operation-type.enum.ts
export enum StoreOperationType {
    None = 0,
    SetSignerSeal = 1,
    SetSubmittion = 2,
    SetSigningType = 3,
    SetDocumentCollectionData = 4,
    SetOtpRequired = 5,
    SetCurrentJumpedField = 6,
    OpenAppendicesFromRemote = 7,
    OpenNotesFromRemote = 8,
    OpenAtthmentsFromRemote = 9,
    SetAttachmentshown = 10,
    SetServerSignatureCredential = 11,
    SetSenderAppendices = 12,
    SetSenderNotes = 13,
    SetOperation = 14,
    SetFieldData = 15,
    SetAttachmentData = 16,
    SetDocumentsData = 17,
    SetselectedSignField = 18,
    SetAttachments = 19,
    SetIsContainServerSignature = 20,
    SetAreAllOtherSignersSigned = 21,
    SetSignerName = 22,
    SetLastSigner = 23,
    SetSignerMeans = 24,
    SetSignerADName = 25,
    SetCompanyLogo = 26,
    CloseSaveSignatureForFutureUse = 27,
    SetOauthNeeded = 28,
    SetOauthDone = 29,
    SetOTPCode = 30,
    SetDocToken = 31,
    SetEIDASSigngingFlow = 32,
}